Fix incorrect function signature
